requirements.txt
torch~=2.7.0+cu126
face_recognition~=1.3.0
numpy~=2.1.2
tqdm~=4.67.1
pillow~=10.2.0
opencv-python~=4.11.0.86
pandas~=2.2.2
requests~=2.32.3
pyyaml~=6.0.2
facenet-pytorch

fastapi~=0.115.12
starlette~=0.46.2